JWT 방식으로 변경

+++ src/main/kotlin/kr/co/shallwecode/module/user/persistence/table/UserTable.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 import org.jetbrains.exposed.sql.Table
 
 object UserTable : Table() {
-    val id = integer("id").autoIncrement()
+    val id = long("id").autoIncrement()
     val email = varchar("email", 100)
     val name = varchar("name", 100)
     val loginId = varchar("login_id", 100).uniqueIndex("user_loginId_unique_index")
@@ -17,7 +17,7 @@ object UserTable : Table() {
 
 @Serializable
 data class UserModel(
-    val id: Int,
+    val id: Long,
     val email: String,
     val name: String,
     val loginId: String,
